// src/core/infrastructure/enum/QuoteType.enum.ts
export enum QuoteVariableType {
    Last = 'Last',
    PctChange = 'PctChange',
    Change = 'Change',
    Volume = 'Volume',
    Turnover = 'Turnover',
}

export enum QuoteStockType {
    GEM = 'gem',
    SB = 'sb',
    ETF = 'etf',
    IGNORE = 'ignore',
    SHC = 'shc',
    SZC = 'szc',
    AH = 'ah',
}

export enum QuoteCategoryType {
    RISE = 'rise',
    DROP = 'drop',
    VOLUME = 'volume',
    TURNOVER = 'turnover',
}